Burger menu changes to black if no header image is present
I really do need to learn about Hugo variable contexts. Split the burger menus into black and white variants because I'm stupid.

This is gonna be tough to keep track of as it currently is. I need to find a way to incorporate the menus as part of site-nav-mobile so I don't need to keep track of when it should be black versus when it should be white.

Also, did some formatting work on single.html. Mostly spacing and stuff