font-weight:300 to fix thin JP description font
This is really just a workaround because it changes all page descriptions to fw300, when I would prefer all the Latin ones to be fw100. It's not awful, it's just not great. Again, really need to play around with Hugo's functions so I can do a comparsion like {{ if eq .Params.description 武士道ボイ }} and then set the font weight using that.