Removed FontAwesome, moved to simple-line-icons
Might need to make the burger menus sit inside inline-block divs
instead of being absolutely positioned, to help with alignment when
opening the menu when viewing a single blog post.

Can't add padding or margins that would make a difference when
absolutely positioned, and I don't want to have to write a CSS class
of like 2.1rem just to get the positioning right.