Remove hugo comment from caddyfile and update readme
Gave the hugo plugin a try this weekend. It isn't great. Filebrowser
is pretty clunky to use, I don't like that it builds it's database
in ~/.caddy, and the new post function didn't work for me.

Really neat idea, and I really wish it would work just so I could write
blog posts on the go, but oh well.